#1044 7. Re-factored dynamic key toString conversion to be more functional and used standardised way of not including key parts: by string representation == '' rather than by value == null. This made conversion of composite entities more consistent with conversion of simple entities in regard of edge-case 'boolean' value 'null': from now composite entities with null 'boolean' value converts to 'false' in all places. This is somewhat artificial example for non-dot-notated properties (because we use 'boolean' instead of 'Boolean' for propety type). However if 'boolean' property is dot-notated (and thus can be null) or Boolean property type is used, we will benefit from such consistent behavior later. (+tests)
