+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/sample/domain/ITgEntityCompositeBooleanKey.java
@@ -1,0 +1,12 @@
+package ua.com.fielden.platform.sample.domain;
+
+import ua.com.fielden.platform.dao.IEntityDao;
+
+/**
+ * Companion object for entity {@link TgEntityCompositeBooleanKey}.
+ *
+ * @author TG Team
+ *
+ */
+public interface ITgEntityCompositeBooleanKey extends IEntityDao<TgEntityCompositeBooleanKey> {
+}
